+++ src/main/java/com/citi/cms/service/impl/UserServiceImpl.java
@@ -27,8 +27,8 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public Optional<User> getUserById(Long id) {
-        return userRepository.findById(id);
+    public Optional<User> getUserById(Integer id) {
+        return userRepository.findById(id.longValue());
     }
 
     @Override
